+++ src/components/OrderButtons.tsx
@@ -7,8 +7,7 @@ import { boundingBoxMatch, combineGranuleFilters } from "../utils/CMR";
 import { IEnvironment } from "../utils/environment";
 import { hasChanged } from "../utils/hasChanged";
 import { UserContext } from "../utils/state";
-import { ConfirmationFlow } from "./ConfirmationFlow";
-import { EarthdataFlow } from "./EarthdataFlow";
+import { EdscFlow } from "./EdscFlow";
 import { EddFlow } from "./EddFlow";
 import { ScriptButton } from "./ScriptButton";
 import { SubmitButton } from "./SubmitButton";
@@ -23,8 +22,7 @@ interface IOrderButtonsProps {
 }
 
 interface IOrderButtonsState {
-  showConfirmationFlow: boolean;
-  showEarthdataFlow: boolean;
+  showEdscFlow: boolean;
   showEddFlow: boolean;
 }
 
@@ -34,8 +32,7 @@ export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButt
   public constructor(props: IOrderButtonsProps) {
     super(props);
     this.state = {
-      showConfirmationFlow: false,
-      showEarthdataFlow: false,
+      showEdscFlow: false,
       showEddFlow: false,
     };
   }
@@ -46,7 +43,7 @@ export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButt
                                                             "orderParameters",
                                                             "orderSubmissionParameters",
                                                             "totalSize"]);
-    const stateChanged = hasChanged(this.state, nextState, ["showConfirmationFlow", "showEarthdataFlow", "showEddFlow"]);
+    const stateChanged = hasChanged(this.state, nextState, ["showEdscFlow", "showEddFlow"]);
 
     return propsChanged || stateChanged;
   }
@@ -94,20 +91,11 @@ export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButt
           tooltip={tooltipEarthdata}
           disabled={orderButtonDisabled}
           onSubmitOrder={this.handleEarthdataOrder} />
-        <ConfirmationFlow
-          cmrGranuleCount={this.props.cmrGranuleCount}
-          environment={this.props.environment}
-          onRequestClose={this.closeConfirmationFlow}
+        <EdscFlow
+          onRequestClose={this.closeEdscFlow}
           onScriptDownloadClick={this.handleScriptDownload}
           orderParameters={this.props.orderParameters}
-          orderSubmissionParameters={this.props.orderSubmissionParameters}
-          show={this.state.showConfirmationFlow}
-          totalSize={this.props.totalSize} />
-        <EarthdataFlow
-          onRequestClose={this.closeEarthdataFlow}
-          onScriptDownloadClick={this.handleScriptDownload}
-          orderParameters={this.props.orderParameters}
-          show={this.state.showEarthdataFlow}
+          show={this.state.showEdscFlow}
           totalSize={this.props.totalSize} />
         <EddFlow
           onRequestClose={this.closeEddFlow}
@@ -119,12 +107,8 @@ export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButt
     );
   }
 
-  private closeConfirmationFlow = () => {
-    this.setState({showConfirmationFlow: false});
-  }
-
-  private closeEarthdataFlow = () => {
-    this.setState({ showEarthdataFlow: false });
+  private closeEdscFlow = () => {
+    this.setState({ showEdscFlow: false });
   }
 
   private closeEddFlow = () => {
@@ -133,7 +117,7 @@ export class OrderButtons extends React.Component<IOrderButtonsProps, IOrderButt
 
   private handleEarthdataOrder = () => {
     this.setState({
-      showEarthdataFlow: true,
+      showEdscFlow: true,
     });
   }