+++ src/renderOrderForm.tsx
@@ -9,7 +9,7 @@ import { EverestUI } from "./components/EverestUI";
 // Render an empty everest-container so that it can be passed to ReactModal
 // before the whole app has rendered. The app element set here should not be a
 // parent of the modal, which is set via ReactModal.parentSelector in
-// ConfirmationFlow. With this setup, #everest-ui is a parent of both
+// e.g., EddFlow. With this setup, #everest-ui is a parent of both
 // div#everest-container and div.ReactModalPortal.
 //
 // http://reactcommunity.org/react-modal/accessibility/#app-element